Add files via upload
Initial check/ validation of the model with observed data sets. And the results in excel file encourage to you the formula provided in the WAPA model since with that in action the total error in power generation seems to be less in comparison to the general power formula used even with 90% turbine efficiency value used.