Latest files with new objective
I am having trouble with the extreme points model(Line  257 onwards). I am not sure why the out put parameter is not caputring the values for both objective. Which means if the model run is with the bug objective then the saving parameter is not capturing the result of the alternative objective but replicating the same results for both objective. I am having hard time finding that even I have reviewed it couple of times.  The gdx file might give you better explanation of the problem i am facing.